move APIFnMakeHash to action_makehash.go
